refactor: clean up the web-cli files
