+++ web-cli/src/components/NamespaceTabs.tsx
