+++ web-cli/src/components/styles/NamespaceTabs.css
