+++ web-cli/src/context/AppContext.tsx
@@ -3,13 +3,11 @@ import { AppState, ScheduledTask } from '../types/types';
 
 const initialState: AppState = {
   value: '',
-  isProcessing: false,
   scheduledTasks: [],
 };
 
 type ActionType =
   | { type: 'SET_VALUE'; payload: string }
-  | { type: 'SET_PROCESSING'; payload: boolean }
   | { type: 'ADD_SCHEDULED_TASK'; payload: ScheduledTask }
   | { type: 'REMOVE_SCHEDULED_TASK'; payload: string }
   | { type: 'CLEAR_VALUE' };
@@ -18,8 +16,6 @@ const appReducer = (state: AppState, action: ActionType): AppState => {
   switch (action.type) {
     case 'SET_VALUE':
       return { ...state, value: action.payload };
-    case 'SET_PROCESSING':
-      return { ...state, isProcessing: action.payload };
     case 'ADD_SCHEDULED_TASK':
       return {
         ...state,
