+++ web-cli/src/services/Api.ts
@@ -1,0 +1,4 @@
+export const API_PORT = process.env.REACT_APP_API_PORT || '3001';
+export const API_BASE_URL = `http://localhost:${API_PORT}/api`;
+
+export const DEFAULT_NAMESPACE = 'orchestrator';
