+++ web-cli/src/services/WorkflowService.ts
@@ -1,9 +1,5 @@
 import { WorkflowResult } from '../types/types';
-
-const API_PORT = process.env.REACT_APP_API_PORT || '3001';
-const API_BASE_URL = `http://localhost:${API_PORT}/api`;
-
-const DEFAULT_NAMESPACE = 'orchestrator';
+import { API_BASE_URL, DEFAULT_NAMESPACE } from './Api';
 
 export async function runWorkflow(
   message: string,
@@ -22,24 +18,11 @@ export async function runWorkflow(
 
     const data = await response.json();
 
-    // If we have an error with status 409, it means a workflow is already running
-    // and the message was scheduled as a task instead
-    if (!response.ok && response.status === 409) {
-      console.log('Workflow already running, task scheduled instead');
-      return {
-        scheduled: true,
-        error: data.error,
-      };
-    }
-
-    // Handle other errors
-    if (!response.ok) {
-      console.error('Error running workflow:', data.error);
-      throw new Error(data.error || 'Unknown error');
-    }
-
-    console.log('Workflow executed successfully:', data.result);
-    return data.result;
+    console.log('Workflow already running, task scheduled instead');
+    return {
+      status: data.status,
+      error: data.error,
+    };
   } catch (error) {
     console.error('Error in runWorkflow:', error);
     throw error;
